% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_bars.R
\name{add_err_bars}
\alias{add_err_bars}
\title{A CLDL Function}
\usage{
add_err_bars(df, plot, measure_var, group_vars)
}
\arguments{
\item{df}{the dataframe (as an object) - use the 'original' dataframe not a recast version, because the summarySE function will recast it}

\item{plot}{the existing ggplot of the data (as an object)}

\item{measure_var}{the name of a column (as a string) that contains the variable to be summarized}

\item{group_vars}{a list containing names of columns (as strings) that contain grouping variables}
}
\value{
returns (1) a png to a 'plots' directory in the current working directory (you should setwd before calling) (2) the new plot to the global environment
}
\description{
This function adds error bars to a ggplot using the summarySE function
}
\examples{
add_err_bars(data2, gleebest_bar, "numResp", c("cond"))

}
\keyword{SE}
\keyword{error}
\keyword{error_bars}
\keyword{ggplot}
\keyword{standard_error}
\keyword{summarySE}

