% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesAnalysis_HelperFunctions.R
\name{allLikelihoods_yepFormat}
\alias{allLikelihoods_yepFormat}
\title{A CLDL Function}
\usage{
allLikelihoods_yepFormat(data, list_id_vars, list_of_hypotheses)
}
\arguments{
\item{data}{the dataframe, **MUST have columns for each hypothesis labeled in the format "[hypothesis]Yep", "[hypothesis]Nope"**}

\item{list_id_vars}{the list of independent variables for the recast function}

\item{list_of_hypotheses}{the list of hypothesis names}
}
\value{
returns dataframe for each hypothesis with likelihoods by ID varibables, and 1 merged dataframe of all hypotheses
}
\description{
This function calculates likelihoods for each hypothesis of a list, for use in our Bayes Analysis
}
\examples{
allLikelihoods_yepFormat(data, c("subject", "mixing", "alignment"), c("strong", "weak", "average", "X1biggest"))

}
\keyword{bayes}
\keyword{hypothesis}
\keyword{likelihood}

