% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_Alternative_Functions.R
\name{calculateKValue}
\alias{calculateKValue}
\title{A CLDL Function}
\usage{
calculateKValue(df, first_hypothesis, rest_hypotheses)
}
\arguments{
\item{df}{the dataframe, must contain the likelihoods for all hypotheses}

\item{first_hypothesis}{the hypothesis for the numerator (to compare to all others)}

\item{rest_hypotheses}{the list of the other hypotheses to compare to}
}
\value{
returns dataframe containing the k values
}
\description{
This function calculates Kvalues for one hypothesis (by comparing its likelihoods to all others), for use in our Bayes Analysis
}
\examples{
calculateKValue(data2, "weak", c("X1biggest", "average"))
calculateKValue(data2, "strong", c("weak", "average", "X1biggest", "X2biggest"))

}
\keyword{bayes}
\keyword{hypothesis}
\keyword{kvalue}

