% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_column.R
\name{correct_column}
\alias{correct_column}
\title{A CLDL Function}
\usage{
correct_column(data, col1, col2, new_dataset)
}
\arguments{
\item{data}{the original dataset}

\item{col1}{the first column name for comparison}

\item{col2}{the second column name for comparison}

\item{new_dataset}{the name for the new dataset, as a string}
}
\value{
returns a new dataset to the global environment with the 'correct' column appended
}
\description{
This function compares values across two specified columns of a dataset, row by row, to create/update values in a 'correct' column. 'correct' is assigned a 1 if the two columns match values for that row, else a 0.
}
\examples{
correct_col(data2, data2$tr2, data2$tr3, "data3")
correct_col(data, data$response, data$expected, "data2")


}
\keyword{0s}
\keyword{1s}
\keyword{column}
\keyword{compare}
\keyword{correct}

