% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesAnalysis_HelperFunctions.R
\name{list_to_recastformula}
\alias{list_to_recastformula}
\title{A CLDL Function}
\usage{
list_to_recastformula(list)
}
\arguments{
\item{list}{the character vector (that will be the id.vars parameter in our recast function)}
}
\value{
returns id.vars as a formula formatted for use in the recast function
}
\description{
This function generates the recast formula from its list of id.vars (primarily called from within other package fxns)
}
\examples{
list_to_recastformula(c("subject", "alignment", "mixing"))

}
\keyword{formula}
\keyword{list}
\keyword{recast}

