% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesAnalysis_HelperFunctions.R
\name{makeNopes_fromYeps}
\alias{makeNopes_fromYeps}
\title{A CLDL Function}
\usage{
makeNopes_fromYeps(df, hypotheses)
}
\arguments{
\item{df}{the dataset; each hypothesis requires a column with subject response consistent with hypothesis = 1, else 0; labeled [hypothesis]Yep}

\item{hypotheses}{the list of hypotheses (list of strings)}
}
\value{
returns complete dataset including nopes to global environment
}
\description{
This function adds, for each hypothesis, a "nope" columns (hypothesis-inconsistent response = 1) based on "yep" columns (hypothesis-consistent response =1), for use in the likelihoods in our Bayes Analysis
}
\examples{
makeNopes_fromYeps(yepdata, c("strong", "weak", "X1biggest", "X2biggest"))

}
\keyword{Bayes}
\keyword{likelihood}
\keyword{nope}
\keyword{yep}

