% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimerge_vertical_csv.R
\name{multimerge_vertical_csv}
\alias{multimerge_vertical_csv}
\title{A CLDL Function}
\usage{
multimerge_vertical_csv(folder_path)
}
\arguments{
\item{folder_path}{path to folder containing individual .csv files, requires they have the same columns}
}
\value{
returns complete dataset to global environment and concatenated/raw csvs to 'Data_preprocessed' directory (created/found in parent dir of folder_path)
}
\description{
This function vertically combines multiple csv datasets and saves dataframe to global environment and a csv to a 'data' folder in your current working directory
}
\examples{
multimerge_vertical_csv("/Users/cec804/Desktop/themostData_6-29-16")
multimerge_vertical_csv("/Users/caseycolby/Desktop/themostData_6-29-16")

}
\keyword{combine}
\keyword{concatenate}
\keyword{csv}
\keyword{merge}
\keyword{post-processing}

