% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json_survey_data.R
\name{parse_json_survey_data}
\alias{parse_json_survey_data}
\title{A CLDL Function}
\usage{
parse_json_survey_data(df, json_colname)
}
\arguments{
\item{df}{the dataframe containing survey data}

\item{json_colname}{{string} the name of the column containing the json objects to parse}
}
\value{
returns dataframe to global environment with columns appended to contain the parsed/separated info from the json object
}
\description{
This function parses the json objects stored in one dataframe column (e.g. JsPsych survey data)
}
\examples{
parse_json_survey_data(surveyData, "responses")

}
\keyword{json}
\keyword{parse}
\keyword{survey}

