% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesAnalysis_HelperFunctions.R
\name{rearrangeSummarize}
\alias{rearrangeSummarize}
\title{A CLDL Function}
\usage{
rearrangeSummarize(df, list_id_vars, list_of_hypotheses)
}
\arguments{
\item{df}{the dataframe, with k values for each hypothesis and summary by hypothesis}

\item{list_id_vars}{the independent variables}

\item{list_of_hypotheses}{the list of hypotheses}
}
\value{
returns dataframe 'data3' with only the columns of interest and a 'winner' column
}
\description{
This function rearranges and summarizes the kvalue data for easier plotting, for use in our Bayes Analysis
}
\examples{
rearrangeSummarize(kdata, c("subject", "mixing", "alignment"), c("strong", "average", "X2biggest"))

}
\keyword{Bayes}
\keyword{KValue}

