% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_long.R
\name{reshape_long}
\alias{reshape_long}
\title{A CLDL Function}
\usage{
reshape_long(filepath, list_of_id_vars, variable_name = NULL,
  value_name = NULL)
}
\arguments{
\item{filepath}{filepath of the dataset}

\item{list_of_id_vars}{variables identifying invididual rows of data}

\item{variable_name}{string: column name for possible variable types}

\item{value_name}{string: column name for the value of those variables}
}
\value{
returns reshaped dataset to global environment
}
\description{
reshapes dataframe from wide(1 subject per line) to long format (1 variable/value per line)
}
\examples{
reshape_long("/Users/caseycolby/Desktop/gleebest2-data.csv", c("subj", "date", "cond", "DOB", "Age", "tr1", "tr2", "tr3"), card_number, response)

}
\keyword{long}
\keyword{melt}
\keyword{reshape}
\keyword{wide}

