% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_type_split.R
\name{trial_type_split}
\alias{trial_type_split}
\title{A CLDL Function}
\usage{
trial_type_split(df, type_list = c("text", "single-stim", "survey-text"))
}
\arguments{
\item{df}{a dataframe in the global environment, or a filepath (string) to a csv}

\item{type_list}{specify a list of trial_type values (as strings) to ouput separate csvs/dataframes for, or leave default; default is set to c("text", "single-"stim", "survey-text")}
}
\value{
returns dataframes corresponding to each specified trial_type (1) to global environment (2) writes .csvs to working directory
}
\description{
This function vertically combines multiple csv datasets and saves dataframe to global environment and a csv to a 'data' folder in your current working directory
}
\examples{
trial_type_split(data)
trial_type_split("~/Desktop/totalData.csv")
trial_type_split(data2, c("text", "single-stim"))
trial_type_split(dataIllu, c("survey-text", "survey-likert"))

}
\keyword{separate}
\keyword{split}
\keyword{survey}
\keyword{trial_type}
\keyword{type}

