% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimerge_vertical_csv.R
\name{multimerge_vertical_csv}
\alias{multimerge_vertical_csv}
\title{A CLDL Function}
\usage{
multimerge_vertical_csv(folder_path, export_path = NULL)
}
\arguments{
\item{folder_path}{path to folder containing individual datasets}

\item{export_path}{= NULL optional path/filename if you want to export the complete csv to that location}
}
\value{
returns complete dataset to global environment
}
\description{
This function vertically combines multiple csv datasets and optionally exports the complete dataset to csv on your machine
}
\examples{
multimerge_vertical_csv("/Users/cec804/Desktop/themostData_6-29-16", "/Users/cec804/Desktop/themost_completeData_6-29-16.csv")
multimerge_vertical_csv("/Users/caseycolby/Desktop/themostData_6-29-16")

}
\keyword{merge}

